cleanup test and create utils
