+++ test/local_e2e/consul_connect_test.go
@@ -4,8 +4,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"io/ioutil"
-	"net"
-	"net/http"
 	"os"
 	"os/exec"
 	"path/filepath"
@@ -18,52 +16,6 @@ import (
 	"github.com/solo-io/gloo/pkg/log"
 )
 
-type ConsulService struct {
-	Service Service `json:"service"`
-}
-
-type Service struct {
-	Name    string  `json:"name"`
-	Port    int     `json:"port"`
-	Connect Connect `json:"connect"`
-}
-
-type Connect struct {
-	Proxy Proxy `json:"proxy"`
-}
-
-type Proxy struct {
-	ExecMode string   `json:"exec_mode"`
-	Command  []string `json:"command"`
-	Config   Config   `json:"config"`
-}
-
-type Config struct {
-	Upstreams []Upstream `json:"upstreams"`
-}
-
-type Upstream struct {
-	DestinationName string `json:"destination_name"`
-	LocalBindPort   int    `json:"local_bind_port"`
-}
-
-type ProxyInfo struct {
-	ProxyServiceID    string
-	TargetServiceID   string
-	TargetServiceName string
-	ContentHash       string
-	ExecMode          string
-	Command           []string
-	Config            ProxyConfig
-}
-
-type ProxyConfig struct {
-	BindAddress         string     `json:"bind_address"`
-	BindPort            uint       `json:"bind_port"`
-	LocalServiceAddress string     `json:"local_service_address"`
-	Upstreams           []Upstream `json:"upstreams"`
-}
-
 var _ = Describe("ConsulConnect", func() {
 	var tmpdir string
 	var consulConfigDir string
@@ -73,7 +25,7 @@ var _ = Describe("ConsulConnect", func() {
 	var consulSession *gexec.Session
 	xdsPort := 7071
 
-	var waitForInit time.Duration = 5 * time.Second
+	var waitForInit time.Duration = 10 * time.Second
 
 	BeforeSuite(func() {
 		var err error
@@ -95,6 +47,7 @@ var _ = Describe("ConsulConnect", func() {
 		gexec.CleanupBuildArtifacts()
 	})
 
+	serviceWritten := false
 	writeService := func(uds bool) {
 		// generate the template
 		args := []string{
@@ -146,6 +99,7 @@ var _ = Describe("ConsulConnect", func() {
 
 		err = ioutil.WriteFile(filepath.Join(consulConfigDir, "service.json"), data, 0644)
 		Expect(err).NotTo(HaveOccurred())
+		serviceWritten = true
 	}
 
 	BeforeEach(func() {
@@ -163,7 +117,7 @@ var _ = Describe("ConsulConnect", func() {
 	})
 
 	AfterEach(func() {
-		gexec.TerminateAndWait("5s")
+		gexec.TerminateAndWait("10s")
 		consulSession = nil
 
 		if tmpdir != "" {
@@ -172,66 +126,33 @@ var _ = Describe("ConsulConnect", func() {
 	})
 
 	runConsul := func() {
-		consul := exec.Command("consul", "agent", "-dev", "--config-dir="+consulConfigDir)
-		session, err := gexec.Start(consul, GinkgoWriter, GinkgoWriter)
-		consulSession = session
-
-		Expect(err).NotTo(HaveOccurred())
+		if !serviceWritten {
+			writeService(false)
+		}
+		consulSession = RunConsul(consulConfigDir)
+	}
+	waitForProxy := func() {
+		// time.Sleep(1 * time.Second)
+		// Expect(consulSession).ShouldNot(gexec.Exit())
+		Eventually(consulSession.Out, "10s").Should(gbytes.Say("agent/proxy: starting proxy:"))
+	}
+	runConsulAndWait := func() {
+		runConsul()
+		waitForProxy()
 	}
 
 	It("should start envoy with gloo tcp", func() {
-		writeService(false)
-		runConsul()
-		time.Sleep(1 * time.Second)
-		Expect(consulSession).ShouldNot(gexec.Exit())
-		Eventually(consulSession.Out, "5s").Should(gbytes.Say("agent/proxy: starting proxy:"))
-
+		runConsulAndWait()
 		// check that a port was opened where consul says it should have been opened (get the port from consul connect and check that it is open)
-		resp, err := http.Get("http://127.0.0.1:8500/v1/agent/connect/proxy/web-proxy")
-		Expect(err).NotTo(HaveOccurred())
-		body, err := ioutil.ReadAll(resp.Body)
-		resp.Body.Close()
-		Expect(err).NotTo(HaveOccurred())
-
-		var cfg ProxyInfo
-		json.Unmarshal(body, &cfg)
-
-		//runFakeXds(cfg.Config.BindAddress, cfg.Config.BindPort)
-
-		time.Sleep(waitForInit)
-
-		conn, err := net.Dial("tcp", fmt.Sprintf("%s:%d", cfg.Config.BindAddress, cfg.Config.BindPort))
-		Expect(err).NotTo(HaveOccurred())
-
-		// We are connected! good enough!
-		conn.Close()
+		cfg := GetProxyInfo()
+		Eventually(func() error { return TestPortOpen(cfg.Config.BindAddress, cfg.Config.BindPort) }, waitForInit, "1s").Should(BeNil())
 	})
 
 	It("should start envoy with gloo uds", func() {
 		writeService(true)
-		runConsul()
-		time.Sleep(1 * time.Second)
-		Expect(consulSession).ShouldNot(gexec.Exit())
-		Eventually(consulSession.Out, "5s").Should(gbytes.Say("agent/proxy: starting proxy:"))
-
-		// check that a port was opened where consul says it should have been opened (get the port from consul connect and check that it is open)
-		resp, err := http.Get("http://127.0.0.1:8500/v1/agent/connect/proxy/web-proxy")
-		Expect(err).NotTo(HaveOccurred())
-		body, err := ioutil.ReadAll(resp.Body)
-		resp.Body.Close()
-		Expect(err).NotTo(HaveOccurred())
-
-		var cfg ProxyInfo
-		json.Unmarshal(body, &cfg)
-
-		//runFakeXds(cfg.Config.BindAddress, cfg.Config.BindPort)
-
-		time.Sleep(waitForInit)
-
-		conn, err := net.Dial("tcp", fmt.Sprintf("%s:%d", cfg.Config.BindAddress, cfg.Config.BindPort))
-		Expect(err).NotTo(HaveOccurred())
-
-		// We are connected! good enough!
-		conn.Close()
+		runConsulAndWait()
+		cfg := GetProxyInfo()
+		Eventually(func() error { return TestPortOpen(cfg.Config.BindAddress, cfg.Config.BindPort) }, waitForInit, "1s").Should(BeNil())
 	})
+
 })
